Fix: Add single source of truth for location change, on map click and on search.
