+++ src/components/TopicPage.svelte
@@ -7,24 +7,10 @@
   import { buildHyperlink } from "../helpers/buildHyperlinkHelper";
   import { selectedGeographyStore } from "../stores/stores";
   import { returnCorrectArticle, unCapitalizeFirstLetter } from "../util/stringUtil";
-  import { fetchGeographyLookup } from "../data/fetchGeographyData";
-  import { getSelectedGeography } from "../helpers/categoryHelpers";
-  import type { GeographyLookupProps } from "../types";
 
   $: topicSlug = $page.params.topic;
   $: topic = topics.find((t) => t.slug === topicSlug);
   $: selectedGeographyDisplayName = $selectedGeographyStore?.displayName;
-  $: search = $page.url.search;
-  $: selectedGeography = getSelectedGeography($page.url);
-
-  $: if (search) {
-    fetchGeographyLookup(selectedGeography.geoCode).then((response) => {
-      const {
-        meta: { name, code, geotype },
-      }: GeographyLookupProps = JSON.parse(response);
-      $selectedGeographyStore = { displayName: name, geoCode: code, geoType: geotype.toLowerCase(), variableData: {} };
-    });
-  }
 </script>
 
 <Heading
