+++ src/data/fetchVizData.ts
@@ -11,13 +11,19 @@ export const fetchVizData = async (args: {
   categoryCode: string;
   categoryCodes: string[];
   geoType: GeoType;
+  geoCode: string;
   bbox: Bbox;
 }) => {
-  const [places, breaks] = await Promise.all([fetchQuery(args), fetchBreaks(args)]);
+  const [places, breaks, table] = await Promise.all([
+    fetchQuery(args),
+    fetchBreaks(args),
+    fetchSelectedGeographyData(args),
+  ]);
   vizStore.set({
     breaks: breaks[args.categoryCode].map((breakpoint) => parseFloat(breakpoint) * 100),
     places: places.map((row) => parsePlaceData(row, args.totalCode, args.categoryCode)),
     params: getCategoryInfo(args.categoryCode),
+    variableData: parseSelectedGeographyData(table, args.totalCode),
   });
   return Promise.resolve();
 };
@@ -41,6 +47,25 @@ const fetchBreaks = async (args: { totalCode: string; categoryCodes: string[]; g
 
   // (ignore data from the API that we don't need)
   return Object.fromEntries(Object.keys(parsed).map((code) => [code, parsed[code][args.geoType.toUpperCase()]]));
+};
+
+const fetchSelectedGeographyData = async (args: { totalCode: string; categoryCodes: string[]; geoCode: string }) => {
+  const url = `${apiBaseUrl}/query/2011?cols=${args.totalCode},${args.categoryCodes.join(",")}&rows=${args.geoCode}`;
+  const response = await fetch(url);
+  const csv = await response.text();
+  return dsv.csvParse(csv);
+};
+
+const parseSelectedGeographyData = (rawData: dsv.DSVRowArray<string>, totalCode: string) => {
+  const total = parseInt(rawData[0][totalCode]);
+  const catCodesArr = rawData.columns.filter((catCode) => catCode != totalCode);
+  const selectedGeographyData = {};
+  catCodesArr.forEach((categoryCode) => {
+    const count = parseInt(rawData[0][categoryCode]);
+    const percentage = (count / total) * 100;
+    selectedGeographyData[categoryCode] = { count: count, total: total, percentage: percentage };
+  });
+  return selectedGeographyData;
 };
 
 const parsePlaceData = (row: dsv.DSVRowString<string>, totalCode: string, categoryCode: string) => {
