+++ src/map/initMap.ts
@@ -2,7 +2,7 @@ import mapboxgl, { Map } from "mapbox-gl";
 import { fromEvent } from "rxjs";
 import { delay, throttleTime } from "rxjs/operators";
 import type { GeoType } from "../types";
-import { vizStore, mapStore } from "../stores/stores";
+import { vizStore, mapStore, selectedGeographyStore } from "../stores/stores";
 import { setGeoSearchParam } from "../helpers/queryParamsHelper";
 import { initMapLayers } from "./initMapLayers";
 import { renderMapViz } from "./renderMapViz";
@@ -19,7 +19,7 @@ export const initMap = (container) => {
   });
 
   map.addControl(new mapboxgl.NavigationControl({ showCompass: false }));
-  
+
   map.on("load", () => {
     initMapLayers(map);
   });
@@ -45,6 +45,12 @@ export const initMap = (container) => {
 
   map.on("click", "msoa-features", (e) => {
     const geoCode = e.features[0].properties["areacd"];
+    const displayName = e.features[0].properties.hclnm;
+    selectedGeographyStore.set({
+      geoType: "msoa",
+      displayName,
+      geoCode,
+    });
     setGeoSearchParam({ geoType: "msoa", geoCode });
   });
